Add missing 'perform_prime' member-variable.
Note that this needs the updated GRPC definitions as well! -- This caused that variable to always be (re)set to false, even when no plugin was present, due to the 'identity' conversion taking place. (Maybe we should stop that in general, since that implies that we copy the data even when no plugin is present.) In any case, even if it wouldn't have caused this particular bug, it's an oversight that should be fixed anyway.

should fix CURA-11084