Don't use voronoi generation in variable xy distance calculations
But instead calculate for each point the closest point on the "other" layer. Unfortunatly this means a nest for loop, looping through all points of the current layer, and for each of those points looping through all points of the other layer. This could mean a decrease in performance, but code is a lot more managable.

CURA-11098