Add optional Sentry integration
The update adds an option to enable or disable Sentry integration in the code. Sentry is configured to capture and report crashes only if the appropriate option is enabled and a valid DSN (data source name) is provided. The suitable changes have been made in the Conan, CMake configurations and source code files.

To compile locally with Sentry support run the conan install command with:
```
-c user.curaengine:sentry_url=<url> -o curaengine:enable_sentry=True
```

For the runners the conf `user.curaengine:sentry_url` has been added to
the relevant runner profiles. Once the new workflows have been merged
that url will move out of the profiles and passed along as a repository
variable in the GH runner.

Contributes to CURA-11364