Add Sentry breadcrumb logging
Integrated Sentry breadcrumbs into CuraEngine through a new sentry_sink
utility. This allows Sentry to collect breadcrumb logs, which can be
useful for debugging. The sentry_sink utility is included in
Application.cpp and is only used when the SENTRY_URL is defined,
ensuring that breadcrumb logging is an opt-in feature.

Contribute to CURA-11482