Optimize wall ordering
Quick optimization we could make; we were checking if one polygon was the child of another polygon. We were doing this for contour parallel tool paths. These tool paths have one property we can exploit, the polygons are never intersecting. A logical consequence from this property is that if one point of the child polygon is inside the parent polygon then all points are inside.

The child-parent check is simplified by chekcing a single point.

CURA-11352