Add emscripten logic
Based on the CuraEngine patch done by Wakeful-Cloud. Not sure if we want
to go the same route but it will give us a quick way forward.
https://github.com/Cloud-CNC/cura-wasm

Doing the conan install with the following profile
```
include(cura.jinja)

[tool_requires]
emsdk/3.1.50

[settings]
os=Emscripten
arch=wasm
```
should do the trick in building it with emsdk. Still very much WIP!

Contributes to NP-5

Co-authored-by: Wakeful-Cloud <wakefulcloud@gmail.com>