Get fractional support from model-overhang instead of layer above.
The detection of where fractional support layers (where the z-distance between model and support was a fractional value compared to support-layer-height) where supposed to be was handled by looking at the internal structure of the support layers itself, rather than where it's actually needed. This fix replaces the previous mehtod, which looked at if the layer above this one had support, and if not, assumed that the gap with the model nust be directly above it (this was originally done both for performance reasons, and making it less dependant on the model and thus more encapsulated) -- but this has proven to be a mistake, given how many problems it introduces.

The new method instead queries the meshes directly for the overhang. Otherwise, any gap between two layers of support would introduce new 'fractional support layers', which aren't supposed to be in those places.

part of CURA-11424