Line-spacing can be 0, causing 0-width grid and div-by 0.
NOTE: This is the lint-formatted version of this change. See the previous commit for the actual change! -- If the combing-offset gets set to 0 (because it's set equal to the line-spacing, which can be 0), we attempt to make a lookup-grid with 0-size squares, causing a division by zero later on. Apply some defensive programming.

Should fix Sentry-issue CURAENGINE/AC.