Fix overlapping lines with gradual (support) infill.
Separate out the areas, since overlapping lines will become a big problem when 'connect infill lines' or 'zig zag' is enabled. This does mean that lines are less well-connected when the pattern _isn't_ already connected in and of itself, like lines. However, such patterns aren't recommended anyway. Also some compensation for the extra density which we _may_ now lose in some cases might need to be done. It should still allow us to to the combine (though I can't see that this happens in practice, either before or after this fix, so we'll may have to look at that in more detail later).

part of CURA-11597