Removed SUPPORT type z seam and added isInDisallowedArea func.
The code changes revolve around the functionality of the InsetOrderOptimizer and related path optimizers. These changes allow for better handling of support meshes by allowing the optimizer to consider if a layer part is supported or not. Furthermore, the type of paths used in path optimization was switched from 'ClipperLib::Paths' to 'Polygons' providing better performance and cleaner code. Also, the conditions for EZSeamType::SUPPORT in the seam configuration were removed.

CURA-11227