use Shape instead of Polygons for the PrimeTowers
This commit replaces instances of the Polygons data structure with Shape in the PrimeTower class, both in prime tower calculations and in storage data structures such as MovesByExtruder and MovesByLayer. This change streamlines the handling of geometric data and enables more robust manipulation of shapes in the prime tower generation process.

Contribute to NP-9830