Use a pointer for the primer tower handler object
The previous versions used an embedded object which was always present
as a member variable. We now instantiate the primer tower object only if
a prime tower will be actually generated, and use this pointer presence
as a test to having a prime tower. This is also in prevision of a future
change where multiple implementations of the prime tower will exist.