Move extruder prime logic calculation to PrimeTower* classes
Now the whole logic of computation of how each extruder should be primed
is done inside the PrimeTowerInterleaved and PrimeTowerNormal classes,
because this is very related to how they will process this data later.
This will later allow for more fine-tuning of these behaviors without a
big risk of breaking each other.