Add cura resources as optional Conan dependency
This update adds an optional dependency to 'cura_resources' in the 'conandata.yml' and 'conanfile.py'. Another significant change is in the 'gcodeanalyzer.yml' where the steps to checkout Cura and the associated environment variables have been removed in favor of adding the Cura resources directly to the Conan install command.

Contribute to NP-186