Update Conan package workflow to run on PR events
This commit modifies the Conan package GitHub action to also be triggered by pull request events, in addition to the existing push events. The action will now run whenever a PR is opened, reopened, or updated, and it's limited to changes in specific paths like 'include/**', 'src/**', etc. on the main branch. This will ensure that the package creation is verified on relevant pull requests.

Contribute to CURA-9830