All printing prime blobs for all extruders when using a raft
Previously we would calculate the most appropriate layer to do the
priming of an extruder, but when using a prime blob the extruder needs
to be primed at first layer, which won't happen when using a raft. We
now then for priming at first layer when having a prime blob, and
process priming after the raft base layer has been printed.

CURA-11868