Force adding temperatures commands when required
Previously, the initial temperatures commands were ignored if the
temperatures were already set in the header. This causes issues because
we don't actually know in what state the printer will let the nozzles
after its init sequence. So we now write non-wait and wait temperatures
commands based only on the settings, so that one can set them properly
according to the printer init sequence.

CURA-11930