Added wasm package creation
Modified conditions for the creation of conan packages in 'conan-package.yml' to exclude 'master' branch. Also added a new package creation workflow for WebAssembly (wasm). This change is temporary for the 'NP-202_conan_wasm_package' branch, needs to be reverted once the related PR is merged to main.

Contribute to NP-202