Fix interaction between prime tower base and support/adhesion
With the new prime tower, the base is now calculated much later on the
process, because it requires to know the extruders uses orders. However,
the interation with support and bed adhesion features is calculated
before that. So we can calculate an approximate outline as soon as the
tower is initialized, and later we calculate the exact toolpaths.