Better reflect current use of function (up'd docs and renamed).
In the current branch (suggested changes to fix the inconsistant small areas w.r.t. innaccurate points of sharp objects), only non-outer walls are removed. This aligns more closely with the originally intended (and at the moment, only) use-case of this function (as specified in the commit where it was added). However, the original function name and documentation/comment, still specified what it used to do, not what it was supposed to, and now does. (Which is removing small lines _from gap filling_ instead of just everywhere.)

part of CURA-9399