Change behavior of infill overlap to match expected behavior
The infill overlap previously worked by extending the whole infill area
and filling it. However, infill areas sometimes have inner walls, and in
this case we would expand the walls, but the pattern generated inside
would not overlap the walls, which is actually what we want to achieve
in order to ensure that the sparser inner pattern properly sticks to
what is outside it, whenever this is an infill wall or a skin wall.