Write the fan speed value only when really different
Previously we were basing the writing or non-writing of the fan speed
command based on the raw value comparison. This could result in some
commands not being written if the new value was too much similar to the
previous one. Now we check the actual output value and base the writing
on this.

CURA-6410