Use push_back instead of emplace_back
Since we can't have the performance improvements provided by
emplace_back and its construct-in-place mechanism without adding a lot
of code (constructors for the struct), just use push_back to emphasize
the fact that we are constructing the objects then putting them into the
list.
This is fine though, we don't actually need performance in this case.