Fuzzy skin: Allow 'inside' to be not just definite 2d holes.
Expand the 'only fuzzy in outside' setting to also mean things the user reasonably could consider inside, instead of just completely closed holes. Take the (approximate) convex hull of the shape, take the difference to get candidate parts that could be considered to be inside of the shape -- then, for each part, determine where it touches the outside. The ratio between how much 'outside' outline and total outline of such a shape (as well as a general size of the hole if ther _is_ any 'outside') determines if points on the polygon thouching that part (outside of the original polygon but inside the approx convex hull) will get 'fuzzed' or not.

part of CURA-11887