Fix possible crash
To compute the end type, we previously used a lambda function, that was
declared static. Thus, the arguments were apparently captured when
calling it the first time, and didn't change for subsequent calls. As
one of them is a pointer, this could crash if the pointed object had
been destroyed in the meantime.
Although the minimum fix is just to remove the static for the function,
I removed the function itself because it really doesn't make sense
anymore in this context.

CURA-12042