The 'spiralize' variable was missing.
In the previous spike, the 'travel_to_z' argument was added (at the end), causing the need for previously omitted (default) parameters to be added as well. In this case, 'spiralize' was omitted, causing the FAN_SPEED_DEFAULT (which is an enum set to -1) to be interpreted as a speed_factor instead.

part of CURA-12081