Fix coasting for 'split-up' paths.
The coasting functionality previously relied (partially) on paths being their own 'complete' extrusions, in between wich where (presumably, mostly) travels. So it only coasted on the last path before a travel, and didn't take into account paths that where to short to complete the full coast (it didn't continue on to the next). This commit fixes that partially -- the volume already coasted isn't taken into account yet, but otherwise it should work as before, but then with the split paths the scarf and accellleration start/end introduced.

part of CURA-12081