Change segment overhanging calculation formula
CURA-12078
Instead of naively checking whether the start and end point of the
segment are in the overhang area, calculate the intersection of the
segment with the overhang area, and see if the length of the
intersection is a significant enough part of the origin segment. This
filters out some segments that partially overhang and are not really
expected to be treated as overhanging.