Uniformize score calculation for seam vertex selection
CURA-12164
We now use a unified score for all the criteria when selecting the best
candidate vertex for the seam. This way we can more easily add new
critera.
At the same time, this scoring is now used to exclude the vertices that
are on overhang areas. We need to do this at the same time as the
regular seam calculation so that it can properly mix with other
criteria.