Change extra infill lines splitting strategy
CURA-12207
The previous strategy was to split the two crossed segments and join
them betweem each other, printing the joined section twice. Although
this is very efficient in terms of path, having a segment printed twice
is sometimes an issue when there is actually some material under the
segment. The new strategy is to create multiple separate paths, which
should create more travel nove but no more segment duplication.