Introduce random scoring in seam candidate calculation
CURA-12164

This way the "random" strategy is also confonted to overhang checking,
so that when having the choice between multiple random points if will
select those not overhanging when possible.
Also partially rewrote the score calculation code to make it very clear
what scores are calculated dependending on the settings. This also
allows doing long pre-calculations only when really required.