Only perform 'tiny loops' fix if there's space to do so.
Previously, we created this 'resolveIntersection' function to handle infill line intersections that where very close to the walls and would therefore create 'tiny loops', which created problems, especially when extruding, like for connect infill lines. However, the alternative maneuvre (or at least the calculation thereof) needs at least _some_ space to work in. If the crossing to be prevented is 'on' the line (within an epsilon), the calculation will not be possible in the current state -- and if things are that close, the rest of the algos will take care of it.

CURA-12173