Apply scarf seam and speed gradient in surface mode
CURA-12176

The addSplitWall method is now a template so that it can handle either
ExtrusionLine or Polygon objects. It also takes a function as parameter
because the way we add extrusion lines is not the same for surface mode
and normal mode.
In order to simplify this function, I added a PathAdapter class to work
with different types of paths with more transparency.