Fix some cases of unaligned seams (e.g. cubes)
CURA-12218
In order to restore the previous seam consistency even better, we now
use two different fallback passes, the first one working on Y only, and
the second one on X only. This way it really mimics the former behavior
and always aligns the seam even on very symmetric shapes.