Fix non-straight user-defined seam
CURA-12264
With a high-definition model, the new vertex created especially for the
seam is quite close to the other points of its former segment. So the
two other vertices would end up at good enough outsiders for the
distance criterion. Now the distance factor of the criterion is a
setting, so that in this specific case we can change the value to give
more weight to the specific vertex and less to others, even if very
close.