Retraction combing avoid distance on 0 would crash.
This isn't something you want, and thankfully doesn't happen in practice, since machines always have a 'machine_nozzle_size' in practice, but our tests take the 'default' in certain cases instead of the 'value' (which can actually be a formula).

Still better to fix the edge-case, or at least work around it a bit more gracefully, regardless of what the tests end up being.