Can have a different number of walls for top/bottom skin.
Note that it was necesary to have the outlines before the wall calculations, so now there's two outline calculation functions.

Also, the current version doesn't subdivide layer-parts when necesary, so if a layer has bot top/bottom parts and parts that aren't, but the piece itself is an undivided polygon, then that whole area will be printed with the top/bottom ammount of walls (will be fixed in subsequent commits).

Laslty (at the moment at least) this doesn't yet take into account gradual changes in the walls, for example, say with about half a line-width -- not sure what the code will do at the moment, but needs to be looked at. (Will also be done in subsequent commits.)

first pass of CURA-12446