Split up parts into top/bottom/side if the wall-count is different.
This commit doesn't yet actually _use_ the wall-difference mind you. -- Note also that the top/bottom has to be calculated here from the rough outlines (as opposed to the actually printed outlines) since normally that would be calculated later. (And due to the previously mentioned print-outlines vs. outlines, it's not the same calculation, so we can't just move it back to here.)

part of CURA-12446