Crude and messy proposal for fix dual-extrusion min-layer time. Don't merge as-is!
Instead of considering the layer-time to be the time spent on a single layer, instead reinterpret it to be the time it takes for an extruder to probably be above the same XY location again. Then, in the case of exactly 2 extruders that have a pattern of switching the order around to minimize the number of extruder switches (like we do), we can say that instead of enforcing minimum-layer time on a single layer with all extruder plans, we instead enforce minimum layer time on the combination of the last extruder-plan of the previous layer, and the first extruder-plan of the current layer.

The current PR is quie crude in that it's just the fastest way to get a prototype out -- it violates DRY a lot, it doesn't contain any considereations for 'exceptional' cases like a single layer with 2 extruders sandwiched between two layers of a single extruder -- it also doesn't even _try_ to adress more than 2 extruers (however, that last bit is less important, as the previous behaviour was only correct for 1 extruder, and sort-of-maybe-workable for > 1 -- now at least it also mostly works for 2).

Ah one last thing Ihad to change is the meaning of 'last' extruder w.r.t. its meaning in the extruder-plan function. Previously it meant the highest extruder-number that was actually used inthe extruder-plans of that layer. Now it means the extruder-number of the (chronologically) latest extruder-plan in that layer.

Needless to say, please don't merge as it is currently -- if you see this in production code, assume that either I or my colleagues have used this as the basis and (heavily?) refactored the current changes to be more up to our standards.

CURA-12460