Fix wrong travel position after extruder switch
CURA-11978
When switching from an extruder to an other one, the current_position_ of the GCodeExport is no more consistent, because this position includes nozzle offset, which depends on the extruder. We now apply the delta of the offsets so that the new position is relative to the new extruder, and subsequent algorithms use the expected position. In this case, calculations over travel move, which now requires the start position.