Fix case where the retract/prime ratio was incorrect
CURA-11978
In a case where e.g. the prime would be 100% done during travel, but we don't have enough time to perform it, we would reduce the prime but not reduce the retract amount during stationary accordingly. Now we retract stationary amounts so that the global retract/prime ratio remains the same.