Hackathon '25: Slice/Print-ID label: Apply label-extents to model/gcode -- but not any ID yet. (WIP)
One the one hand, a lot got done, on the other, this commit is about 90% FIXME's an TODO's by weight -- What is done here is that you can see that it applies a gcode offset where the label is to be printed (for the slice ID), and it alters the gcode (with a comment) as well (for future use in the firmware hopefully). It's just that at the moment, no actual ID is 'printed' so it doesn't function as a slice-ID yet, and the gcode may be commented, but not with the UV's from the ID-label (not to be confused with the paint-texture) so it can't function as a print-ID yet. -- Furthermore, I think I only got the projection 'working' on one side of the cube, and even then it projects it scrimped in the horizontal direction on the opposit side -- to say nothing of the IdFieldInfo properties that've been set up in the previous commit (and should be altered anyway) but that isn't even properly used yet either.

part of UMH-2025