Rewrite 'sliced UV coords' to use _line_ grid instead of _point_ grid.
This gets rid of the awkward '99% of cases then handle the 1% with a different algorithm' thing,
_but more importantly_ it sets the stage for properly using that class (in the next commit(s) hopefully) to get 'lines' of UV-coordinates that match up with the lines across face(s).
The previous approach was adequate if you just wanted to know of a corner what UV-coords it had and didn't care which face it 'choose', because the paint would overlap the entire corner anyway (such as with the seam).
However, since the UV-textures may be disconnected along the face-borders, and the points we're using _likely match up with the face corners_, this repressents a problem if you have to have a line across a face,
as the function can only return one of the (maximum of) two possitbilities (an edge is part of at most 2 faces), and currently does so arbitrary -- I mean it still does kinda, but like I said, stage set for next steps.

part of (Hackathon '25 project Slice-ID/Print-ID) UMH-2025