Optimize computation time by using 2 lookup trees
CURA-12587
It turns out the lookup time to find the closest point in the tree increases exponentially when the closest point is further from the position, so as we are iterating to move deeper inside the mesh, the lookup times become crazy. So now we have a second tree with a very low resolution, which is used when we are deep enough inside the mesh.