Fix closed shaped being completely filled
CURA-12770
TL;DR: wibbly-wobbly geometric stuff, problem fixed
This was due to dropping the holes when making the final horizontal polygons. Although it makes sense for closed polygons, it does not for those who actually have a hole inside.
Now the propagation strategy has been changed by checking the insideness at voxel evaluation time, so that voxels considered outside the mesh will also be considered for the next evaluation. In the end, this creates full closed shells around the painted features, so that we can later distinguish the inside of the mesh (which contains unassigned voxels) from the outside (which contains voxels assigned with the mesh extruder), so now we can create properly closed polygons where they should be.