Fix blobs between painted areas
CURA-12772
When making the polygons for 2 additional extruders, some of their edges perfectly overlap. So when later applying them on the mesh, rounding errors can leave a tiny space in between, producing blobs of the main extruders to fill the area. We now apply a small offset to make sure the areas properly overlap.