Refactor computeDistanceToBridgeStart and call it only when necessary
CURA-11989
This function is quite heavy because it loops over the whole extrusion line and processes intersection calculations. So now, instead of calling it for every processed segment, call it at first then just decrease the distance of each processed segment, until we actually reach bridging area, then recalculate it.
Also the function didn't take care of the direction in which the extrusion line was being processed, and whether it could be closed.